+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `setDevicePin` GraphQL mutation which lets the `owner` of a `Device` or an `admin`set its `pin` (or anyone if the `Device` has not been claimed).
 * `setDeviceName` GraphQL mutation which lets the `owner` of a `Device` or an `admin` set its `name`.
 * `loginDevice` GraphQL query for logging in a `Device` using `mac` and `pin`.
+* `txBeacon` GraphQL mutation which is used by `Device` for showing its presence online.
 
 ### Changed
 * `User` mongoose model requires `firstName` and `lastName`.